requirements.txt
torch
torchvision
Pillow
opencv-python
numpy
matplotlib  # For visualization (optional, but helpful)
# Add other dependencies here as you go (e.g., for GANs, pose estimation)absl-py==2.3.1
attrs==25.3.0
cffi==1.17.1
config==0.5.1
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.7.0
jax==0.7.0
jaxlib==0.7.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.16.1
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
